Implement reconnect logic
This will retry to connect on certain actions. It will try up until
x times and has a exponential backoff (with a max of 10 seconds).

It will retry when an issue occurs during connecting while increasing
the tries. Once a connection is established, the current try will reset
to zero again. If an connection is dropped without sending the Close
frame then the Proxy will also retry.

Also removed the reunite and close logic as this is taken care of by the
tungstenite library.

If an error occured while running, proxy will now return an exit code 1.